+++ src/app/home-module/services/edit.service.ts
@@ -29,4 +29,16 @@ export class EditService extends BehaviorSubject<Student[]> {
     };
     super.next(this.data);
   }
+
+  public ageCalculation(birthday: Date): number {
+    const today = new Date();
+    const birthDate = new Date(birthday);
+    let age = today.getFullYear() - birthDate.getFullYear();
+    const m = today.getMonth() - birthDate.getMonth();
+    if (m < 0 || (m === 0 && today.getDate() < birthDate.getDate())) {
+      age--;
+    }
+
+    return age;
+  }
 }
